Refactor code for querying API
