--- PackagesOfFuture.WebUI/src/app/models/payment-type.ts
@@ -1,4 +1,0 @@
-export class PaymentType {
-    id!: number; 
-    description!: string;
-}
