+++ PackagesOfFuture.WebUI/src/app/models/payments.ts
@@ -1,0 +1,4 @@
+export interface PaymentType {
+  id: number; 
+  description: string;
+}
