--- PackagesOfFuture.WebUI/src/app/models/service-model.ts
@@ -1,5 +1,0 @@
-export class ServiceModel {
-    name!: string;
-    description!: string;
-    price!: number;
-}
