+++ PackagesOfFuture.WebUI/src/app/models/services.ts
@@ -1,0 +1,5 @@
+export interface ServiceModel {
+  name: string;
+  description: string;
+  price: number;
+}
