+++ PackagesOfFuture.WebUI/src/app/models/users.ts
@@ -1,0 +1,5 @@
+export interface ChangePasswordDto {
+  userId: number, 
+  oldPassword: string, 
+  newPassword: string
+}
